+++ behave_django/__init__.py
@@ -1,4 +1,4 @@
 """Behave BDD integration for Django"""
 
-__version__ = '0.4.0'
+__version__ = '0.4.1'
 __license__ = 'MIT License'